use_cases/time_series_prediction/Time_Series_SRC/requirements.txt
dtw_python==1.3.0
matplotlib==3.7.1
numpy==1.24.3
pandas==2.0.1
sdv==1.1.0
torch==1.13.1
